4.16 done
Add a feature which adds the following restrictions to creating new users: Both username and password must be given. Both username and password must be at least 3 characters long. The username must be unique.

The operation must respond with a suitable status code and some kind of an error message if an invalid user is created.

Also, implement tests that ensure invalid users are not created and that an invalid add user operation returns a suitable status code and error message.