6.10 and 6.11 done
Install Redux Toolkit for the project. Move the Redux store creation into the file `store.js` and use Redux Toolkit's `configureStore` to create the store.

Change the definition of the `filter reducer and action creators` to use the Redux Toolkit's `createSlice` function.

Change also the definition of the `anecdote reducer and action creators` to use the Redux Toolkit's `createSlice` function.