6.12 and 6.13 done
Extend the ready-made `Notification` component so that it renders the message stored in the Redux store.

Make changes to the application's existing reducer. Create a separate reducer for the new functionality by using the Redux Toolkit's `createSlice` function.

Extend the application so that it uses the Notification component to display a message for five seconds when the user votes for an anecdote or creates a new anecdote.

Create separate action creators for setting and removing notifications.