6.16 - 6.18 done
Modify the initialization of the Redux store to happen using asynchronous action creators, which are made possible by the Redux Thunk library.

Also modify the creation of a new anecdote to happen using asynchronous action creators, made possible by the Redux Thunk library.

Fix the voting with the help of the Redux Thunk library.