7.1 done
Add React Router to the application so that by clicking links in the Menu component the view can be changed.

At the root of the application, meaning the path /, show the list of anecdotes

The Footer component should always be visible at the bottom.

The creation of a new anecdote should happen e.g. in the path create