7.3 done
Improve the `create new` functionality such that after creating a new anecdote the application transitions automatically to showing the view for all anecdotes and the user is shown a notification informing them of this successful creation for the next five seconds.

Also modify the application so that the Anecdote component receives only the anecdote that it should display.