7.8 done
Extract the code for communicating with the backend into its own `useResource` hook.

Implement fetching all resources and creating a new resource.

The `useResource` custom hook returns an array of two items just like the state hooks. The first item of the array contains all of the individual resources and the second item of the array is an object that can be used for manipulating the resource collection, like creating new ones.