7.14 and 7.15 done
Implement a view to the application that displays all of the basic information related to users.

Implement a view for individual user that displays all of the blog posts added by that user.

One can access the view by clicking the name of the user in the view that lists all users.